// src/shared/enums/roles.enum.ts
/**
 * @description This enum represents the roles of the user. Only used on the AuthGuard
 * @author Damian
 * @date 13/06/2025
 * @export
 * @enum {number}
 */
export enum RolesEnum {
  ADMINISTRATOR = 'administrator',
  AUTHOR = 'author',
  CONTRIBUTOR = 'contributor',
  CSS_JS_DESIGNER = 'css_js_designer',
  EDITOR = 'editor',
  EXAMENES = 'examenes',
  EXAMENESV2 = 'examenesv2',
  EXAMENESV3 = 'examenesv3',
  NO_SOCIO = 'no_socio',
  PRUEBA = 'prueba',
  SOCIO = 'socio',
  SUBSCRIBER = 'subscriber',
  TRANSLATOR = 'translator',
}